% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ro_show_colors.R
\name{get_color_palette}
\alias{get_color_palette}
\title{Return a color palette}
\usage{
get_color_palette(palette_name, custom_colors = NULL)
}
\arguments{
\item{palette_name}{A string specifying the name of the palette. Valid options
are "full", "categorical", "gender_con", "gender_unc", "greys", and "custom".}

\item{custom_colors}{An optional character vector of hexadecimal color codes.
This parameter is required if \code{palette_name} is "custom".}
}
\value{
A named character vector of hexadecimal color codes representing the selected color palette.
}
\description{
This function returns a predefined color palette based on the specified palette name.
Custom color palettes can also be provided using hexadecimal color codes. If \code{custom_colors}
is unnamed, this function names the colors by position (from 1 to \code{length(custom_colors)}).
}
\examples{
# Examples with a predefined palette
(ROvis.utils:::get_color_palette("full"))
(ROvis.utils:::get_color_palette("gender_unc"))

# Example with a custom palette using hexcodes directly
(ROvis.utils:::get_color_palette("custom", c("#FF5733", "#33FF57", "#3357FF")))

# Example using 'ro_color()'
(ROvis.utils:::get_color_palette("custom", ro_color_seq(ro_color("robijnrood_tint15"),
ro_color("robijnrood"), 12)))

}
\seealso{
Other ggplot2:
\code{\link[=ro_check_if_font_available]{ro_check_if_font_available()}},
\code{\link[=ro_color]{ro_color()}},
\code{\link[=ro_color_categorical]{ro_color_categorical()}},
\code{\link[=ro_color_palette]{ro_color_palette()}},
\code{\link[=ro_color_seq]{ro_color_seq()}},
\code{\link[=ro_color_seq_named]{ro_color_seq_named()}},
\code{\link[=ro_color_validate]{ro_color_validate()}},
\code{\link[=ro_get_color_palette]{ro_get_color_palette()}},
\code{\link[=ro_gg_create_palette]{ro_gg_create_palette()}},
\code{\link[=ro_show_colors]{ro_show_colors()}},
\code{\link[=ro_show_colors_get_df]{ro_show_colors_get_df()}}
}
\concept{ggplot2}
\keyword{internal}
