% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ro_show_colors.R
\name{ro_gg_create_palette}
\alias{ro_gg_create_palette}
\title{Create a tile plot of a palette with ggplot}
\usage{
ro_gg_create_palette(df, palette_name, palette)
}
\arguments{
\item{df}{dataframe with 3 character columns: number, name and hex}

\item{palette_name}{Character. Name of the palette to use in plot title}

\item{palette}{named character vector with hex values to use in scale_fill_manual}
}
\value{
ggplot object
}
\description{
Create a tile plot of a palette with ggplot
}
\examples{
\dontrun{
palette_name <- "categorical"
palette <- ROvis.utils:::get_color_palette(palette_name)
df <- ROvis.utils:::ro_show_colors_get_df(palette)
ROvis.utils:::ro_gg_create_palette(df, palette_name, palette)
}
}
\seealso{
Other ggplot2:
\code{\link[=get_color_palette]{get_color_palette()}},
\code{\link[=ro_check_if_font_available]{ro_check_if_font_available()}},
\code{\link[=ro_color]{ro_color()}},
\code{\link[=ro_color_categorical]{ro_color_categorical()}},
\code{\link[=ro_color_palette]{ro_color_palette()}},
\code{\link[=ro_color_seq]{ro_color_seq()}},
\code{\link[=ro_color_seq_named]{ro_color_seq_named()}},
\code{\link[=ro_color_validate]{ro_color_validate()}},
\code{\link[=ro_get_color_palette]{ro_get_color_palette()}},
\code{\link[=ro_show_colors]{ro_show_colors()}},
\code{\link[=ro_show_colors_get_df]{ro_show_colors_get_df()}}
}
\concept{ggplot2}
\keyword{internal}
