% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ro_show_colors.R
\name{ro_show_colors_get_df}
\alias{ro_show_colors_get_df}
\title{Create a data frame from a color palette}
\usage{
ro_show_colors_get_df(palette)
}
\arguments{
\item{palette}{A named character vector of hex color codes.}
}
\value{
A tibble (data frame) with columns \code{number}, \code{name} and \code{hex}.
}
\description{
This function takes a color palette and returns a data frame containing the
color names and their corresponding hex values. If the palette has named colors,
the data frame will include these names and ensure the correct order.
If the palette is unnamed, it will generate a sequence of numbers to label the colors.
}
\examples{
ROvis.utils:::ro_show_colors_get_df(c(red = "#FF0000", green = "#00FF00", blue = "#0000FF"))
ROvis.utils:::ro_show_colors_get_df(c("1" = ro_color("donkerblauw"), "2" = ro_color("robijnrood"),
 "3" = ro_color("mintgroen")))
}
\seealso{
Other ggplot2:
\code{\link[=get_color_palette]{get_color_palette()}},
\code{\link[=ro_check_if_font_available]{ro_check_if_font_available()}},
\code{\link[=ro_color]{ro_color()}},
\code{\link[=ro_color_categorical]{ro_color_categorical()}},
\code{\link[=ro_color_palette]{ro_color_palette()}},
\code{\link[=ro_color_seq]{ro_color_seq()}},
\code{\link[=ro_color_seq_named]{ro_color_seq_named()}},
\code{\link[=ro_color_validate]{ro_color_validate()}},
\code{\link[=ro_get_color_palette]{ro_get_color_palette()}},
\code{\link[=ro_gg_create_palette]{ro_gg_create_palette()}},
\code{\link[=ro_show_colors]{ro_show_colors()}}
}
\concept{ggplot2}
\keyword{internal}
