% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROvis.utils-package.R
\docType{package}
\name{ROvis.utils-package}
\alias{ROvis.utils}
\alias{ROvis.utils-package}
\title{ROvis.utils: Utilities for Rijksoverheid-Styled Data Visualizations}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides shared utilities for the ROvis suite, including Rijksoverheid color palettes, palette validation, and font-availability checks, so that themes across the ROvis packages stay consistent with official Rijksoverheid design guidelines.
}
\author{
\strong{Maintainer}: ROvis team \email{spin@rivm.nl}

Authors:
\itemize{
  \item ROvis team \email{spin@rivm.nl}
}

}
\keyword{internal}
