% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROvis.utils-package.R
\docType{package}
\name{ROvis.utils-package}
\alias{ROvis.utils}
\alias{ROvis.utils-package}
\title{ROvis.utils: Rijksoverheid Visualisatie - utils}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

ROvis.utils is an R package that provides a comprehensive suite of utilities for visualizing in Rijksoverheid style.
}
\author{
\strong{Maintainer}: ROvis team \email{spin@rivm.nl}

Authors:
\itemize{
  \item ROvis team \email{spin@rivm.nl}
}

}
\keyword{internal}
